requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
annotated-types
anyio
attrs
certifi
cffi
charset-normalizer
dataclasses-json
distro
faiss-cpu
frozenlist
greenlet
h11
httpcore
httpx
httpx-sse
idna
iniconfig
jiter
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-ollama
langchain-openai
langchain-pinecone
langchain-tests
langchain-text-splitters
langsmith
markdown-it-py
marshmallow
mdurl
multidict
mypy_extensions
numpy
ollama
openai
orjson
packaging
pinecone
pinecone-plugin-assistant
pinecone-plugin-interface
pluggy
propcache
py-cpuinfo
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
pypdf
PyPDF2
pytest
pytest-asyncio
pytest-benchmark
pytest-codspeed
pytest-recording
pytest-socket
python-dateutil
python-dotenv
PyYAML
regex
requests
requests-toolbelt
rich
six
sniffio
SQLAlchemy
syrupy
tenacity
tiktoken
tqdm
typing-inspect
typing-inspection
typing_extensions
urllib3
vcrpy
wrapt
yarl
zstandard